ref: move curves instances to curve library
