+++ lib/crypto/src/curve/sw/instance/mod.rs
@@ -1,0 +1,3 @@
+//! This module contains curve instances in the Short Weierstrass model.
+pub mod secp256k1;
+pub mod starknet;
