+++ lib/crypto/src/curve/te/instance/mod.rs
@@ -1,0 +1,2 @@
+//! This module contains curve instances in the Twisted Edwards model.
+pub mod curve25519;
